+++ ving/docs/change-log.md
@@ -7,6 +7,7 @@ outline: deep
 
 ### 2024-06-06
 * Fixed documentation for ving/utils/fs readJSON and writeJSON
+* Added an optional function that will be passed to an array filter to filter out any unwanted records from the current list of records on useVingKind() records list.
 
 ### 2024-06-01
 * Documented noSetAll() attribute in Ving Schemas.